Confirming file downloaded contains a certificate
Before you could send unlimited notifications to bouncy with random URLs on valid AWS SNS domains, causing bouncy to download and store in the cache every file even if it failed. This ran the risk of filling the cache. This will prevent that by only allowing valid PEM files containing an actual certificate to be stored in the cache.